// frontend/src/api/models/bulkCategoryActionResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { BulkCategoryActionResponseSkippedItem } from './bulkCategoryActionResponseSkippedItem'

export interface BulkCategoryActionResponse {
  /** @minimum 0 */
  requested: number
  /** @minimum 0 */
  updated: number
  skipped: BulkCategoryActionResponseSkippedItem[]
}